requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
Werkzeug==3.1.3

Flask-Migrate~=4.1.0
alembic~=1.16.5